src/main/java/org/languagetool/oxygen/Helper.java
/* LanguageTool plugin for Oxygen XML editor
 * Copyright (C) 2015 Daniel Naber (http://www.danielnaber.de)
 *
 * This library is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 2.1 of the License, or (at your option) any later version.
 *
 * This library is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this library; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301
 * USA
 */
package org.languagetool.oxygen;

import java.util.ArrayList;
import java.util.List;

final class Helper {

  private Helper() {
  }

  static List<String> splitAtSpace(String s, int max) {
    List<String> result = new ArrayList<String>();
    StringBuilder sb = new StringBuilder();
    for (int i = 0; i < s.length(); i++) {
      char c = s.charAt(i);
      if (sb.length() >= max && c == ' ') {
        result.add(sb.toString());
        sb = new StringBuilder();
      } else {
        sb.append(c);
      }
    }
    if (sb.length() > 0) {
      result.add(sb.toString());
    }
    return result;
  }


}
